move worker functions to internal package
